more seperation between x/rollapp and genesisBridge
