+++ proto/dymensionxyz/dymension/rollapp/rollapp.proto
@@ -9,6 +9,7 @@ import "cosmos/msg/v1/msg.proto";
 
 import "dymensionxyz/dymension/rollapp/state_info.proto";
 import "dymensionxyz/dymension/rollapp/metadata.proto";
+import "dymensionxyz/dymension/rollapp/genesis_info.proto";
 
 // RollappGenesisState is a partial repr of the state the hub can expect the
 // rollapp to be in upon genesis
@@ -77,25 +78,6 @@ message Rollapp {
   int64 last_state_update_height = 18;
 }
 
-message GenesisInfo {
-  // checksum used to verify integrity of the genesis file
-  string genesis_checksum = 1;
-  // unique bech32 prefix
-  string bech32_prefix = 2;
-  // native_denom is the base denom for the native token
-  DenomMetadata native_denom = 3 [ (gogoproto.nullable) = false ];
-  // initial_supply is the initial supply of the native token
-  string initial_supply = 4 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable) = false
-  ];
-
-  // sealed indicates if the fields in this object are no longer updatable
-  bool sealed = 5;
-
-  repeated GenesisAccount genesis_accounts = 6 [ (gogoproto.nullable) = false ];
-}
-
 // Rollapp summary is a compact representation of Rollapp
 message RollappSummary {
   // The unique identifier of the rollapp chain.
@@ -107,17 +89,4 @@ message RollappSummary {
   StateInfoIndex latestFinalizedStateIndex = 3;
   uint64 latestHeight = 4;          // TODO:
   uint64 latestFinalizedHeight = 5; // TODO:
-}
-
-
-
-// GenesisAccount is a struct for the genesis account for the rollapp
-message GenesisAccount {
-  // amount of coins to be sent to the genesis address
-  string amount = 1 [
-    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
-    (gogoproto.nullable) = false
-  ];
-  // address is a bech-32 address of the genesis account
-  string address = 2;
 }
